Add voting functionality
Note: QuestionDetails component makes another request upon successful vote to re-fetch the whole question.

I first considered taking the voting response and updating the choices array of the question object already in memory. But overall, this approach seemed fragile, so I decided to take the performance hit in favor of simplicity and robustness.

I think this is an indication that this app in production would require more powerful state management, e.g. Redux.